refactor: make FollowButton work for artists too
