+++ lib/followers/isUserFollowing.ts
@@ -1,8 +1,11 @@
 import getSpotifyData from "@/lib/getSpotifyData";
 
-const isUserFollowing = async (username: string) =>
+const isUserFollowing = async (
+  usernameOrId: string,
+  followType: "user" | "artist"
+) =>
   getSpotifyData<SpotifyApi.UserFollowsUsersOrArtistsResponse>(
-    `https://api.spotify.com/v1/me/following/contains?type=user&ids=${username}`,
+    `https://api.spotify.com/v1/me/following/contains?type=${followType}&ids=${usernameOrId}`,
     {
       cache: "no-cache",
     }
