+++ lib/followers/toggleFollowing.ts
@@ -2,9 +2,13 @@
 
 import putSpotifyData from "@/lib/putSpotifyData";
 
-const toggleFollowing = async (username: string, isFollowing: boolean) =>
+const toggleFollowing = async (
+  username: string,
+  isFollowing: boolean,
+  followType: "user" | "artist"
+) =>
   putSpotifyData(
-    `https://api.spotify.com/v1/me/following?type=user&ids=${username}`,
+    `https://api.spotify.com/v1/me/following?type=${followType}&ids=${username}`,
     {
       method: isFollowing ? "DELETE" : "PUT",
       cache: "no-cache",
